backend/requirements.txt
django
djangorestframework
markdown
django-filter
drf-yasg